Added empty array as fallback for settings.auto_discover_settings.
When the config is cached and the `settings.auto_discover_settings` setting is missing. It returns `null`.
This produces the error ` Spatie\LaravelSettings\Support\DiscoverSettings::within(): Argument #1 ($directories) must be of type array, null given`.

The within function of DiscoverSettings only accepts an array, so I added an array as the default value.

it is also possible to change the function to `within(array $directories = [])`. I don't know what is better.